update packages and fix tests to work in jest@24
