+++ e2e/test/test.setup.js
@@ -1,7 +1,5 @@
 'use strict';
 
-require('jest-extended');
-
 // We need long timeouts for some of these jobs
 jest.setTimeout(3 * 60 * 1000);
