+++ package.json
@@ -40,16 +40,16 @@
         "typescript": "^3.3.3"
     },
     "devDependencies": {
-        "@types/jest": "^24.0.5",
-        "eslint": "^5.12.1",
+        "@types/jest": "^24.0.6",
+        "eslint": "^5.14.1",
         "eslint-config-airbnb-base": "^13.1.0",
         "eslint-plugin-import": "^2.16.0",
         "fs-extra": "^7.0.1",
-        "jest": "^23.6.0",
+        "jest": "^24.1.0",
         "jest-extended": "^0.11.1",
         "lerna-alias": "^3.0.2",
         "semver": "^5.6.0",
-        "ts-jest": "^23.10.5",
+        "ts-jest": "^24.0.0",
         "tslint": "^5.12.1",
         "tslint-config-airbnb": "^5.11.1"
     },
