+++ packages/teraslice-cli/package.json
@@ -31,8 +31,8 @@
         "archiver": "^3.0.0",
         "bluebird": "^3.5.3",
         "chalk": "^2.4.2",
-        "easy-table": "^1.1.1",
         "cli-table3": "^0.5.1",
+        "easy-table": "^1.1.1",
         "fs-extra": "^7.0.1",
         "lodash": "^4.17.11",
         "node-yaml": "^3.1.1",
@@ -41,7 +41,7 @@
         "teraslice-client-js": "^0.7.0",
         "tmp": "^0.0.33",
         "tty-table": "^2.7.0",
-        "yargs": "^12.0.5"
+        "yargs": "^13.2.1"
     },
     "devDependencies": {
         "jest-fixtures": "^0.6.0"
