+++ packages/teraslice-cli/test/test.setup.js
@@ -1,5 +1,3 @@
 'use strict';
 
-require('jest-extended');
-
 process.env.TJM_TEST_MODE = 'true';
