+++ packages/teraslice/package.json
@@ -53,16 +53,16 @@
         "kubernetes-client": "^5.4.0",
         "lodash": "^4.17.11",
         "moment": "^2.24.0",
+        "nanoid": "^2.0.1",
         "p-whilst": "^1.0.0",
         "porty": "^3.1.1",
         "request": "^2.88.0",
         "shortid": "^2.2.13",
-        "nanoid": "^2.0.1",
         "socket.io": "^1.7.4",
         "socket.io-client": "^1.7.4",
         "terafoundation": "^0.7.2",
         "uuid": "^3.3.2",
-        "yargs": "^12.0.5"
+        "yargs": "^13.2.1"
     },
     "devDependencies": {
         "@terascope/teraslice-op-test-harness": "^1.5.3",
