+++ packages/teraslice/test/test.setup.js
@@ -1,7 +1,5 @@
 'use strict';
 
-require('jest-extended');
-
 process.env.BLUEBIRD_LONG_STACK_TRACES = '1';
 process.env.USE_DEBUG_LOGGER = 'true';
