+++ packages/ts-transforms/package.json
@@ -49,13 +49,13 @@
         "valid-url": "^1.0.9",
         "validator": "^10.10.0",
         "xlucene-evaluator": "^0.5.0",
-        "yargs": "^12.0.5"
+        "yargs": "^13.2.1"
     },
     "devDependencies": {
         "@types/lodash": "^4.14.121",
         "@types/valid-url": "^1.0.2",
         "@types/validator": "^10.9.0",
-        "@types/yargs": "^12.0.8"
+        "@types/yargs": "^12.0.9"
     },
     "engines": {
         "node": ">=8.0.0"
